Use pytest directly
