+++ .github/actions/run-component-tests/action.yaml
@@ -20,18 +20,18 @@ runs:
   steps:
     - name: "Prepare Python environment"
       shell: bash
-      run: python -m virtualenv roles/oneagent/tests/component/venv && 
-           source roles/oneagent/tests/component/venv/bin/activate &&
-           pip install -r roles/oneagent/tests/component/resources/requirements.txt
+      run: |
+        python -m virtualenv venv
+        source venv/bin/activate
+        pip install -r roles/oneagent/tests/requirements.txt
 
     - name: "Install the collection"
       shell: bash
-      run: sudo bash -c "source roles/oneagent/tests/component/venv/bin/activate && ansible-galaxy collection install -vvv dynatrace-oneagent*"
+      run: sudo bash -c "source venv/bin/activate && ansible-galaxy collection install -vvv dynatrace-oneagent*"
 
     - name: "Run component tests"
       shell: bash
-      run: cd roles/oneagent/tests/component && 
-           sudo bash -c "source venv/bin/activate && python run.py --linux_x86=localhost"
+      run: sudo bash -c "source venv/bin/activate && pytest roles/oneagent/tests --linux_x86=localhost"
 
     - name: "Upload logs"
       uses: actions/upload-artifact@v4
@@ -39,4 +39,4 @@ runs:
       with:
         name: component-tests-${{ github.sha }}
         retention-days: 1
-        path: roles/oneagent/tests/component/test_dir/logs
+        path: test_dir/logs
