+++ roles/oneagent/tests/resources/ansible/credentials.yml
