+++ roles/oneagent/tests/resources/ansible/oneagent.yml
