+++ roles/oneagent/tests/resources/installers/oneagentctl.sh
