+++ roles/oneagent/tests/resources/installers/uninstall.sh
