+++ roles/oneagent/tests/technology/constants.py
