+++ roles/oneagent/tests/technology/util.py
