+++ roles/oneagent/tests/test_localInstaller.py
