+++ roles/oneagent/tests/util/constants/common_constants.py
@@ -4,10 +4,12 @@ from pathlib import Path
 # TODO: is cwd() correct?
 COMPONENT_TEST_BASE = Path().cwd() / "test_dir"
 TEST_DIRECTORY = COMPONENT_TEST_BASE / "working_dir"
-RESOURCES_DIRECTORY = Path().cwd() / "resources"
+RESOURCES_DIRECTORY = Path(__file__).resolve().parent.parent.parent / "resources"
 INSTALLERS_DIRECTORY = COMPONENT_TEST_BASE / "installers"
 SERVER_DIRECTORY = COMPONENT_TEST_BASE / "server"
 INSTALLERS_RESOURCE_DIR = RESOURCES_DIRECTORY / "installers"
+LOG_DIRECTORY = COMPONENT_TEST_BASE / "logs"
+
 INSTALLER_CERTIFICATE_FILE_NAME = "dt-root.cert.pem"
 INSTALLER_PRIVATE_KEY_FILE_NAME = "dt-root.key"
 SERVER_CERTIFICATE_FILE_NAME = "server.pem"
